# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación API ",
        "description": "Documentación API ",
        "contact": {
            "name": "API Support",
            "email": "tommycontreras34@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email Usuario",
                                        "type": "string",
                                        "example": "tommy112@gmail.com"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "example": "Hola1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "Create a new user",
                "description": "Sign in",
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "username": {
                                                "type": "string"
                                            },
                                            "first_name": {
                                                "type": "string"
                                            },
                                            "last_name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "telephone": {
                                                "type": "string"
                                            },
                                            "age": {
                                                "type": "number"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "first_name": "Tommy",
                                    "last_name": "Grullón Contreras",
                                    "email": "tommy@gmail.com",
                                    "telephone": "829-754-6150",
                                    "age": 20,
                                    "password": "Hola1234"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "first_name": {
                                            "type": "string",
                                            "example": "Tommy"
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": "Grullón Contreras"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "tommy@gmail.com"
                                        },
                                        "telephone": {
                                            "type": "string",
                                            "example": "829-754-6150"
                                        },
                                        "age": {
                                            "type": "number",
                                            "example": 20
                                        },
                                        "password": {
                                            "type": "string",
                                            "example": "Hola1234"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/profile": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Profile",
                "operationId": "Profile",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Logout",
                "operationId": "Logout",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    }
}